with nodejs, gql client (mock-fe) subscription parameter type is Int which maps to the schema type in mock-gql

+++ mock-fe/src/App.js
@@ -52,7 +52,7 @@ const client = new ApolloClient({
 });
 
 const COMMENTS_SUBSCRIPTION = gql`
-  subscription Subscription($id: Float) {
+  subscription Subscription($id: Int) {
     getStatus(ID: $id) {
       id
       status
